backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pypdf
openai
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
twilio
